% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{check_geometry_type}
\alias{check_geometry_type}
\title{check sf geometry type}
\usage{
check_geometry_type(sfj)
}
\arguments{
\item{sfj}{}
}
\value{
A character
}
\description{
check sf geometry type
}
\examples{
data("income")
check_geometry_type(income)

}
\keyword{internal}
