% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{check_wt}
\alias{check_wt}
\title{check whether wt is a matrix class}
\usage{
check_wt(wt)
}
\arguments{
\item{wt}{Spatial weight matrix}
}
\value{
A \code{matrix}.
}
\description{
If \code{wt} is a \code{matrix} class, return \code{wt} itself, othe rwise an error is raised
and execution stops.
}
\examples{
data("income")
wt = sdsfun::inverse_distance_swm(income)
wt = check_wt(wt)
wt[1:5,1:5]

}
\keyword{internal}
