% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoc_swm.R
\name{geoc_swm}
\alias{geoc_swm}
\title{spatial weight matrix based on geographical complexity}
\usage{
geoc_swm(gc, wt, style = "W")
}
\arguments{
\item{gc}{Geographical complexity vector.}

\item{wt}{Spatial weight matrix based on spatial adjacency or spatial distance relationships.}

\item{style}{(optional) A character that can be \code{B},\code{W},\code{C}.  More to see \code{spdep::nb2mat()}.
Default is \code{W}.}
}
\value{
A matrix
}
\description{
spatial weight matrix based on geographical complexity
}
\examples{
data("income")
inc = dplyr::select(income,Income)
gc = inc \%>\%
  geocd_vector(returnsf = FALSE) \%>\%
  dplyr::pull(1)
wt = sdsfun::spdep_contiguity_swm(inc,style = 'B')
wt_gc = geoc_swm(gc,wt)
wt_gc[1:5,1:5]

}
