% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoc_raster.R
\name{geocd_raster}
\alias{geocd_raster}
\title{calculates geocomplexity for raster data based on spatial dependence}
\usage{
geocd_raster(r, order = 1, normalize = TRUE)
}
\arguments{
\item{r}{Raster object that can be converted to \code{SpatRaster} by \code{terra::rast()}.}

\item{order}{(optional) The order of the adjacency object. Default is \code{1}.}

\item{normalize}{(optional) Whether to further normalizes the spatial local complexity.
Default is \code{TRUE}.}
}
\value{
A SpatRaster object
}
\description{
This function calculates geocomplexity, a geospatial local complexity indicator,
for variable in raster data based on spatial dependence.
}
\examples{
library(terra)
m = matrix(c(3,3,3,3,1,3,
             3,3,3,2,1,2,
             3,3,3,1,2,1,
             1,3,2,2,2,2,
             2,2,2,1,1,2,
             1,2,1,1,1,1),
           nrow = 6,
           byrow = TRUE)
m = rast(m)
names(m) = 'sim'
plot(m, col = c("#d2eaac", "#a3dae1", "#8cc1e1"))
gc1 = geocd_raster(m,1)
gc2 = geocd_raster(m,2)
gc1
gc2

}
