% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoch_raster.R
\name{geoch_raster}
\alias{geoch_raster}
\title{geocomplexity for raster data based on spatial heterogeneity}
\usage{
geoch_raster(r, order = 1, normalize = TRUE, method = "spvar")
}
\arguments{
\item{r}{Raster object that can be converted to \code{SpatRaster} by \code{terra::rast()}.}

\item{order}{(optional) The order of the adjacency object. Default is \code{1}.}

\item{normalize}{(optional) Whether to further normalizes the calculated geocomplexity.
Default is \code{TRUE}.}

\item{method}{(optional) When \code{method} is \code{spvar}, variation of the attribute vector is
represented using spatial variance, otherwise shannon information entropy is used. Default
is \code{spvar}.}
}
\value{
A SpatRaster object
}
\description{
This function calculates geocomplexity for raster data based on spatial heterogeneity.
}
\examples{
library(terra)
m = matrix(c(3,3,3,3,1,3,
             3,3,3,2,1,2,
             3,3,3,1,2,1,
             1,3,2,2,2,2,
             2,2,2,1,1,2,
             1,2,1,1,1,1),
           nrow = 6,
           byrow = TRUE)
m = rast(m)
names(m) = 'sim'
plot(m, col = c("#d2eaac", "#a3dae1", "#8cc1e1"))
gc1 = geoch_raster(m,1)
gc2 = geoch_raster(m,2)
gc1
gc2

}
