% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoch_vector.R
\name{geoch_vector}
\alias{geoch_vector}
\title{geocomplexity for vector data based on spatial stratified heterogeneity}
\usage{
geoch_vector(sfj, wt = NULL, normalize = TRUE, method = "spvar")
}
\arguments{
\item{sfj}{An \code{sf} object or vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{wt}{(optional) Spatial weight matrix. Must be a \code{matrix} class. You can get a
spatial weight matrix from \code{spdep},\code{rgeoda} or \code{tidyrgeoda} package. If \code{wt} is not
provided, \code{geocomplexity} will use a first-order queen adjacency binary matrix via
\code{spdep} package.}

\item{normalize}{(optional) Whether to further normalizes the calculated geocomplexity.
Default is \code{TRUE}.}

\item{method}{(optional) When \code{method} is \code{spvar}, variation of the attribute vector is
represented using spatial variance, otherwise shannon information entropy is used. Default
is \code{spvar}.}
}
\value{
An sf object
}
\description{
This function calculates geocomplexity for in vector data based on spatial stratified heterogeneity.
}
\examples{
data("income")
inc = dplyr::select(income,Income)
gc = geoch_vector(inc)
gc

library(ggplot2)
library(viridis)
ggplot(gc) +
   geom_sf(aes(fill = Geocomplexity_Income)) +
   scale_fill_viridis(option="mako", direction = -1) +
   theme_bw()

}
