% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocs_raster.R
\name{geocs_raster}
\alias{geocs_raster}
\title{calculates geocomplexity for raster data based on geographical similarity}
\usage{
geocs_raster(r, order = 1, normalize = TRUE)
}
\arguments{
\item{r}{Raster object that can be converted to \code{SpatRaster} by \code{terra::rast()}.}

\item{order}{(optional) The order of the adjacency object. Default is \code{1}.}

\item{normalize}{(optional) Whether to further normalizes the calculated geocomplexity.
Default is \code{TRUE}.}
}
\value{
A SpatRaster object
}
\description{
This function calculates geocomplexity for raster data based on geographical similarity.
}
\examples{
library(terra)
m1 = matrix(c(3,3,3,3,1,3,
              3,3,3,2,1,2,
              3,3,3,1,2,1,
              1,3,2,2,2,2,
              2,2,2,1,1,2,
              1,2,1,1,1,1),
           nrow = 6,
           byrow = TRUE)
m1 = rast(m1)
names(m1) = 'sim1'
m2 = m1
set.seed(123456789)
values(m2) = values(m1) + runif(ncell(m1),-1,1)
names(m2) = 'sim2'
m = c(m1,m2)
gc1 = geocs_raster(m,1)
gc2 = geocs_raster(m,2)
gc1
gc2

}
