% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income}
\alias{income}
\title{Spatial datasets of income.}
\format{
\code{income}: An sf object about income
}
\usage{
income
}
\description{
Spatial datasets of income.
}
\author{
Yongze Song \email{yongze.song@curtin.edu.au}
}
\keyword{dataset}
\keyword{income}
