% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocomplexity_helper.R
\name{inverse_distance_weight}
\alias{inverse_distance_weight}
\title{calculate inverse distance weight}
\usage{
inverse_distance_weight(sfj, power = 1)
}
\arguments{
\item{sfj}{Vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{power}{(optional) Default is 1. Set to 2 for gravity weights.}
}
\value{
A inverse distance weight matrices with class of \code{matrix}.
}
\description{
Function for calculate inverse distance weight.
}
\details{
The inverse distance weight formula is
\eqn{w_{ij} = 1 / d_{ij}^\alpha}
}
\examples{
data("income")
wt = inverse_distance_weight(income)
wt[1:5,1:5]

}
\author{
Wenbo Lv \email{lyu.geosocial@gmail.com}
}
