% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocomplexity_helper.R
\name{normalize_vector}
\alias{normalize_vector}
\title{normalization}
\usage{
normalize_vector(x)
}
\arguments{
\item{x}{A continuous numeric vector.}
}
\value{
A continuous vector which has normalized.
}
\description{
normalization
}
\examples{
normalize_vector(c(-5,1,5,0.01,0.99))

}
