% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_geoc.R
\name{print.gcgwrm}
\alias{print.gcgwrm}
\title{print GCGWR result}
\usage{
\method{print}{gcgwrm}(x, ...)
}
\arguments{
\item{x}{Return by \code{gwr_geoc()}.}

\item{...}{(optional) Other arguments.}
}
\value{
Formatted string output
}
\description{
print GCGWR result
}
