% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranI_test.R
\name{print.moran_test}
\alias{print.moran_test}
\title{print global spatial autocorrelation test result}
\usage{
\method{print}{moran_test}(x)
}
\description{
print global spatial autocorrelation test result
}
