% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{st_geometry_name}
\alias{st_geometry_name}
\title{get the geometry column name of an sf object}
\usage{
st_geometry_name(sfj)
}
\arguments{
\item{sfj}{An \code{sf} object.}
}
\value{
A character.
}
\description{
get the geometry column name of an sf object
}
\examples{
data("income")
st_geometry_name(income)

}
