% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{standardize_vector}
\alias{standardize_vector}
\title{standardization}
\usage{
standardize_vector(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A standardized numeric vector
}
\description{
To calculate the Z-score using variance normalization, the formula is as follows:

\eqn{Z = (x - mean(x)) / sd(x)}
}
\examples{
standardize_vector(1:10)

}
