% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocomplexity_helper.R
\name{check_wt}
\alias{check_wt}
\title{check whether wt is a matrix class}
\usage{
check_wt(wt)
}
\arguments{
\item{wt}{Spatial weight matrix}
}
\value{
If \code{wt} is a \code{matrix} class, return \code{wt} itself, otherwise an error is raised
and execution stops.
}
\description{
check whether wt is a matrix class
}
\examples{
data("income")
wt = inverse_distance_weight(income)
check_wt(wt)

}
