% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoc_spatraster.R
\name{geoc_raster}
\alias{geoc_raster}
\title{calculates geocomplexity for raster data}
\usage{
geoc_raster(r)
}
\arguments{
\item{r}{Raster data that can be converted to \code{SpatRaster} by \code{terra::rast()}.}
}
\value{
A SpatRaster object
}
\description{
This function calculates geocomplexity, a geospatial local complexity indicator,
for variable in raster data. The resulting output is A SpatRaster object.
}
\examples{

library(terra)

m = matrix(
  c(3,3,3,3,1,3,
    3,3,3,2,1,2,
    3,3,3,1,2,1,
    1,3,2,2,2,2,
    2,2,2,1,1,2,
    1,2,1,1,1,1),
  nrow = 6,
  byrow = TRUE
)

m = rast(m)
plot(m, col = c("#d2eaac", "#a3dae1", "#8cc1e1"))
a = geoc_raster(m)
a

}
