% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocd_vector.R
\name{geocd_vector}
\alias{geocd_vector}
\title{geocomplexity for vector data based on spatial dependence}
\usage{
geocd_vector(sfj, wt = NULL, normalize = TRUE)
}
\arguments{
\item{sfj}{An \code{sf} object or vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{wt}{(optional) Spatial weight matrix. Must be a \code{matrix} class. You can get a
spatial weight matrix from \code{spdep},\code{rgeoda} or \code{tidyrgeoda} package. If \code{wt} is not
provided, \code{geocomplexity} will use a first-order queen adjacency binary matrix via
\code{spdep} package.}

\item{normalize}{(optional) Whether to further normalizes the calculated geocomplexity.
Default is \code{TRUE}.}
}
\value{
An sf object
}
\description{
This function calculates geocomplexity for vector data based on spatial dependence.
}
\details{
The formula is

\eqn{\rho_i = -\frac{1}{m} Z_i \sum\limits_{j=1}^m W_{ij} Z_j -\frac{1}{m} \sum\limits_{j=1}^m W_{ij} Z_j \frac{1}{V_{k}}\sum\limits_{k=1}^n W_{jk} W_{ik} Z_k}
}
\examples{
data("income")
inc = dplyr::select(income,Income)
gc = geocd_vector(inc)
gc

library(ggplot2)
library(viridis)
ggplot(gc) +
   geom_sf(aes(fill = Geocomplexity_Income)) +
   scale_fill_viridis(option="mako", direction = -1) +
   theme_bw()

}
\references{
Zehua Zhang, Yongze Song, Peng Luo & Peng Wu (2023) Geocomplexity explains spatial errors,
International Journal of Geographical Information Science, 37:7, 1449-1469,
DOI: 10.1080/13658816.2023.2203212

Anselin, L. (2019). A local indicator of multivariate spatial association: Extending
geary’s c. Geographical Analysis, 51(2), 133–150. https://doi.org/10.1111/gean.12164
}
