% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocd_vector.R
\name{geocd_vector}
\alias{geocd_vector}
\title{calculates geocomplexity for vector data based on spatial dependence}
\usage{
geocd_vector(sfj, wt = NULL, normalize = TRUE)
}
\arguments{
\item{sfj}{Vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{wt}{(optional) Spatial weight matrix. Must be a \code{matrix} class. You can get a
spatial weight matrix from \code{spdep},\code{rgeoda} or \code{tidyrgeoda} package. If \code{wt} is not
provided, \code{geocomplexity} will use a first-order queen adjacency binary matrix.}

\item{normalize}{(optional) Whether to further normalizes the spatial local complexity.
Default is \code{TRUE}.}
}
\value{
An sf object
}
\description{
This function calculates geocomplexity, a geospatial local complexity indicator,
for variable in vector data based on spatial dependence.
}
\examples{
library(sf)
data("income")
inc = dplyr::select(income,Income)
gc = geocd_vector(inc)
gc

}
