% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocd_vector.R
\name{geocd_vector}
\alias{geocd_vector}
\title{geocomplexity for vector data based on spatial dependence}
\usage{
geocd_vector(
  sfj,
  wt = NULL,
  method = "moran",
  normalize = TRUE,
  returnsf = TRUE
)
}
\arguments{
\item{sfj}{An \code{sf} object or vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{wt}{(optional) Spatial weight matrix. Must be a \code{matrix} class.}

\item{method}{(optional) In instances where the method is \code{moran}, geocomplexity is
determined using local moran measure method. Conversely, when the method is \code{spvar},
the spatial variance of attribute data serves to characterize geocomplexity. For all
other methods, the shannon information entropy of attribute data is employed to represent
geocomplexity. The selection of the method can be made from any one of the three options:
\code{moran}, \code{spvar} or \code{entropy}. Default is \code{moran}.}

\item{normalize}{(optional) Whether to further normalizes the calculated geocomplexity.
Default is \code{TRUE}.}

\item{returnsf}{(optional) When \code{returnsf} is \code{TRUE}, return an sf object, otherwise a tibble.
Default is \code{TRUE}.}
}
\value{
A tibble (\code{returnsf} is \code{FALSE}) or an sf object (\code{returnsf} is \code{TRUE})
}
\description{
This function calculates geocomplexity for vector data based on spatial dependence.
}
\details{
The formula for geocomplexity which uses local moran measure method is

\eqn{\rho_i = -\frac{1}{m} Z_i \sum\limits_{j=1}^m W_{ij} Z_j -\frac{1}{m} \sum\limits_{j=1}^m W_{ij} Z_j \frac{1}{V_{k}}\sum\limits_{k=1}^n W_{jk} W_{ik} Z_k}
}
\note{
If \code{wt} is not provided, for polygon vector data, \code{geocomplexity} will use a first-order queen
adjacency binary matrix; for point vector data, the six nearest points are used as adjacency
objects to generate an adjacency binary matrix.
}
\examples{
data("income")
inc = dplyr::select(income,Income)
gc = geocd_vector(inc)
gc

library(ggplot2)
library(viridis)
ggplot(gc) +
   geom_sf(aes(fill = GC_Income)) +
   scale_fill_viridis(option="mako", direction = -1) +
   theme_bw()

}
\references{
Zehua Zhang, Yongze Song, Peng Luo & Peng Wu (2023) Geocomplexity explains spatial errors,
International Journal of Geographical Information Science, 37:7, 1449-1469,
DOI: 10.1080/13658816.2023.2203212

Anselin, L. (2019). A local indicator of multivariate spatial association: Extending
geary’s c. Geographical Analysis, 51(2), 133–150. https://doi.org/10.1111/gean.12164
}
