% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocs_vector.R
\name{geocs_vector}
\alias{geocs_vector}
\title{calculates geocomplexity for vector data based on spatial similarity}
\usage{
geocs_vector(sfj, wt = NULL, normalize = TRUE)
}
\arguments{
\item{sfj}{Vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{wt}{(optional) Spatial weight matrix. Must be a \code{matrix} class. You can get a
spatial weight matrix from \code{spdep},\code{rgeoda} or \code{tidyrgeoda} package. If \code{wt} is not
provided, \code{geocomplexity} will use a first-order inverse distance weight matrix via
\code{inverse_distance_weight()}.}

\item{normalize}{(optional) Whether to further normalizes the spatial local complexity.
Default is \code{TRUE}.}
}
\value{
An sf object
}
\description{
This function calculates geocomplexity for in vector data based on spatial similarity.
}
\examples{
library(sf)
data("income")
income = dplyr::select(income,-SA3_CODE16)
gc = geocs_vector(income)
gc

library(ggplot2)
library(viridis)
ggplot(gc) +
   geom_sf(aes(fill = Geocomplexity_Income_Gini)) +
   scale_fill_viridis(option="mako", direction = -1) +
   theme_bw()

}
