% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocs_vector.R
\name{geocs_vector}
\alias{geocs_vector}
\title{geocomplexity for vector data based on geographical similarity}
\usage{
geocs_vector(
  sfj,
  wt = NULL,
  normalize = TRUE,
  similarity = 1,
  method = "spvar"
)
}
\arguments{
\item{sfj}{An \code{sf} object or vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{wt}{(optional) Spatial weight matrix. Must be a \code{matrix} class. You can get a
spatial weight matrix from \code{spdep},\code{rgeoda} or \code{tidyrgeoda} package. If \code{wt} is not
provided, \code{geocomplexity} will use a first-order inverse distance weight matrix via
\code{inverse_distance_weight()}.}

\item{normalize}{(optional) Whether to further normalizes the calculated geocomplexity.
Default is \code{TRUE}.}

\item{similarity}{(optional) When \code{similarity} is \code{1}, the similarity is calculated using
geographical configuration similarity, otherwise the cosine similarity is calculated.
Default is \code{1}.}

\item{method}{(optional) When \code{method} is \code{spvar}, variation of the similarity vector is
represented using spatial variance, otherwise shannon information entropy is used. Default
is \code{spvar}.}
}
\value{
An sf object
}
\description{
This function calculates geocomplexity for in vector data based on geographical similarity.
}
\examples{
data("income")
income = dplyr::select(income,-SA3_CODE16)
gc = geocs_vector(income)
gc

library(ggplot2)
library(viridis)
ggplot(gc) +
   geom_sf(aes(fill = Geocomplexity_Income_Gini)) +
   scale_fill_viridis(option="mako", direction = -1) +
   theme_bw()

}
