% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocs_vector.R
\name{geocs_vector}
\alias{geocs_vector}
\title{geocomplexity for vector data based on geographical similarity}
\usage{
geocs_vector(
  sfj,
  wt = NULL,
  method = "spvar",
  similarity = 1,
  normalize = TRUE,
  returnsf = TRUE
)
}
\arguments{
\item{sfj}{An \code{sf} object or vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{wt}{(optional) Spatial weight matrix. Must be a \code{matrix} class. If \code{wt} is not
provided, \code{geocomplexity} will use a first-order inverse distance weight matrix via
\code{sdsfun::inverse_distance_swm()} function.}

\item{method}{(optional) When \code{method} is \code{spvar}, variation of the similarity vector is
represented using spatial variance, otherwise shannon information entropy is used. Default
is \code{spvar}.}

\item{similarity}{(optional) When \code{similarity} is \code{1}, the similarity is calculated using
geographical configuration similarity, otherwise the cosine similarity is calculated.
Default is \code{1}.}

\item{normalize}{(optional) Whether to further normalizes the calculated geocomplexity.
Default is \code{TRUE}.}

\item{returnsf}{(optional) When \code{returnsf} is \code{TRUE}, return an sf object, otherwise a tibble.
Default is \code{TRUE}.}
}
\value{
A tibble (\code{returnsf} is \code{FALSE}) or an sf object (\code{returnsf} is \code{TRUE})
}
\description{
This function calculates geocomplexity for in vector data based on geographical similarity.
}
\details{
The geographical configuration similarity is calculated as:

\eqn{S({\bf u}_{\alpha},{\bf v}_{\beta})=min\{E_{i}(e_{i}({\bf u}_{\alpha}),e_{i}({\bf v}_{\beta}))\}}

\eqn{E_{i}({\bf u}_{\alpha},{\bf v}_{\beta})=\exp\left(-{\frac{\left(e_{i}({\bf u}_{\alpha})-e_{i}({\bf v}_{\beta})\right)^{2}}{2\left(\sigma^{2}/\delta({\bf v}_{\beta})\right)^{2}}}\right)}

\eqn{\delta({\bf u}_{\alpha},{\bf v})=\sqrt{\frac{\sum_{\beta=1}^{n}(e({\bf u}_{\alpha})-e({\bf v}_{\beta}))^{2}}{n}}}

The spatial variance is calculated as:

\eqn{\Gamma = \frac{\sum_i \sum_{j \neq i} \omega_{ij}\frac{(y_i-y_j)^2}{2}}{\sum_i \sum_{j \neq i} \omega_{ij}}}
}
\examples{
data("income")
income = dplyr::select(income,-SA3_CODE16)
gc = geocs_vector(income)
gc

library(ggplot2)
library(viridis)
ggplot(gc) +
   geom_sf(aes(fill = Income_Gini_geoc)) +
   scale_fill_viridis(option="mako", direction = -1) +
   theme_bw()

}
