% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranI_test.R
\name{moran_test}
\alias{moran_test}
\title{global spatial autocorrelation test}
\usage{
moran_test(sfj, wt = NULL, alternative = "greater", symmetrize = FALSE)
}
\arguments{
\item{sfj}{An \code{sf} object or vector object that can be converted to \code{sf} by \code{sf::st_as_sf()}.}

\item{wt}{(optional) Spatial weight matrix. Must be a \code{matrix} class. You can get a
spatial weight matrix from \code{spdep},\code{rgeoda} or \code{tidyrgeoda} package. If \code{wt} is not
provided, \code{geocomplexity} will use a first-order queen adjacency binary matrix via
\code{spdep} package.}

\item{alternative}{(optional) Specification of alternative hypothesis as 'greater' (default),
'lower', or 'two.sided'.}

\item{symmetrize}{(optional) Whether or not to symmetrize the asymmetrical spatial weight matrix
\emph{\strong{wt}} by: 1/2 * (\emph{\strong{wt}} + \emph{\strong{wt}}'). Default is \code{FALSE}.}
}
\value{
A lis with \code{moran_test} class.
}
\description{
Spatial autocorrelation test based on global Moran'I Index.
}
\note{
This is a \verb{C++} implementation of the \code{MI.vec} function in \code{spfilteR} package,
and embellishes the console output.

The return result of this function is actually a \code{list}, please access the result
tibble using \verb{$result}.

The non-numeric columns of the attribute columns in \code{sfj} are ignored.
}
\examples{
data("income")
moran_test(income)

}
